pyproject.toml
[project]
name = "vigil"
version = "0.1.0"
description = "NAS video downloader via Telegram"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.2",
    "yt-dlp>=2025.1.26",
    "python-telegram-bot>=20.8",
    "apscheduler>=3.10.4",
    "loguru>=0.7.2",
    "pydantic-settings>=2.2.1",
    "sqlalchemy>=2.0.28",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.3.0",
    "bandit>=1.7.7",
]
